requirements.txt
Babel==1.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.8
django-appconf==1.0.1
django-bootstrap-form==3.2
django-bower==5.0.4
django-compressor==2.0
django-extra-views==0.6.4
django-haystack==2.3.2
django-oscar==1.1.1
django-registration-redux==1.2
django-tables2==1.0.7
django-treebeard==4.0
django-widget-tweaks==1.3
factory-boy==2.4.1
gunicorn==19.4.3
mock==1.0.1
phonenumbers==6.3.0
Pillow==2.7.0
psycopg2==2.6.1
purl==1.1
pycountry==1.19
pysolr==3.3.3
pytz==2015.7
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.2
static3==0.6.1
Unidecode==0.4.18
uWSGI==2.0.12
wheel==0.24.0
whitenoise==2.0.6
